Update to regolith3
While updating I also fixed that I had copied over the configuration
files from regolith. Instead I am now just configuring things through
Xresources. The advantage of this is that things will not get broken in
the future. The disadvantage is that the keybindings displayed by
regolith are not actually correct, because we are not modifying the
comments in the config files like we could if we customised them.