Update backup script
Added a whole lot of features.
Preventing the user from shutting down during backup as much as possible.
Detecting locks and if they exist, ask the user to break them or not.
Logging when the last successful backup happened and showing it to the user.
Using trap command to prevent creating locks in the first place as much
as possible.