Update keyboard key service
Not sure if that is due to me or the build of this keyboard or the way
the laptop repair installed the keyboard, but the scancodes for Fn and
the Ctrl keys are now reversed from what they usually are. Also if I
press the left Ctrl Key+K, it maps K to a different Key (noticed through
Emacs, can be observed with xev).