Add scratchpad mode
I am not super happy with the idea of having longer keybinds to open
things in the scratchpad, on the other hand I am running out of keybinds
in i3 due to using so many scratchpad windows. I think I might just
remove some of the less used ones and keep the ones most in heavy use
for emacs, chatgpt and claude.