Add org-protocol file
The org-protocol file had been incorrect (because it assumed ~/bin was
in the path which it is neither in firefox nor chrome leading to a long
search for where it was defined. So I am adding it to the version
control here to document any further important changes that I clearly
need to document, since I didn't get it right on a first try.