Switch back from chromium to chrome
While chrome is annyoing to install with nix on ubuntu, (Just manually
installed it with apt), it is less of a annoyance than trying to install
addons on chromium. So I decided to switch back after one of my favorite
addons was difficult to handle on chromium and others are probably worse.