Remove ssh option again
First I added this option, because borg backup was complaining about it.
I then learned it was not complaining about it being set true, it was
not recognizing the option at all (perhaps a spelling error, but could
also be an old ssh version on remote).
I then changed that option to "no" under /etc/ssh/ssh_config instead,
which worked.